refactor(fingerprint): move fingerprint dep-info to a module
